state management through ModX
