+++ package.json
@@ -10,6 +10,9 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@babel/plugin-proposal-decorators": "^7.12.13",
+    "mobx": "^6.1.7",
+    "mobx-react": "^7.1.0",
     "react": "16.13.1",
     "react-native": "0.63.4"
   },
@@ -18,6 +21,7 @@
     "@babel/runtime": "7.12.13",
     "@react-native-community/eslint-config": "1.1.0",
     "babel-jest": "25.5.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.5",
     "eslint": "6.8.0",
     "jest": "25.5.4",
     "metro-react-native-babel-preset": "0.59.0",
